+++ solaris/vector/mask.py
@@ -1,4 +1,4 @@
-from ..utils.core import _check_df_load, _check_geom
+from ..utils.core import _check_df_load, _check_geom, _check_crs
 from ..utils.core import _check_skimage_im_load, _check_rasterio_im_load
 from ..utils.geo import gdf_get_projection_unit, reproject
 from ..utils.geo import geometries_internal_intersection
@@ -13,7 +13,6 @@ import rasterio
 from rasterio import features
 from affine import Affine
 from skimage.morphology import square, erosion, dilation
-
 
 def df_to_px_mask(df, channels=['footprint'], out_file=None, reference_im=None,
                   geom_col='geometry', do_transform=None, affine_obj=None,
@@ -611,7 +610,7 @@ def buffer_df_geoms(df, buffer, meters=False, reference_im=None,
         reference_im = _check_rasterio_im_load(reference_im)
 
     if hasattr(df, 'crs'):
-        orig_crs = df.crs
+        orig_crs = _check_crs(df.crs)
     else:
         orig_crs = None  # will represent pixel crs
 
@@ -648,10 +647,10 @@ def buffer_df_geoms(df, buffer, meters=False, reference_im=None,
         lambda x: x.buffer(buffer))
 
     # return to original crs
-    if getattr(df_for_buffer, 'crs', None) != orig_crs:
+    if _check_crs(getattr(df_for_buffer, 'crs', None)) != orig_crs:
         if orig_crs is not None and \
                 getattr(df_for_buffer, 'crs', None) is not None:
-            buffered_df = df_for_buffer.to_crs(orig_crs)
+            buffered_df = df_for_buffer.to_crs(orig_crs.to_wkt())
         elif orig_crs is None:  # but df_for_buffer has one: meters=True case
             buffered_df = geojson_to_px_gdf(df_for_buffer, reference_im)
         else:  # orig_crs exists, but df_for_buffer doesn't have one
@@ -782,7 +781,7 @@ def mask_to_poly_geojson(pred_arr, channel_scaling=None, reference_im=None,
             ref.close()
     else:
         transform = Affine(1, 0, 0, 0, 1, 0)  # identity transform
-        crs = None
+        crs = rasterio.crs.CRS()
 
     mask = mask_arr > bg_threshold
     mask = mask.astype('uint8')
@@ -799,7 +798,7 @@ def mask_to_poly_geojson(pred_arr, channel_scaling=None, reference_im=None,
             values.append(value)
 
     polygon_gdf = gpd.GeoDataFrame({'geometry': polygons, 'value': values},
-                                   crs=crs)
+                                   crs=crs.to_wkt())
     if simplify:
         polygon_gdf['geometry'] = polygon_gdf['geometry'].apply(
             lambda x: x.simplify(tolerance=tolerance)
@@ -945,6 +944,16 @@ def instance_mask(df, out_file=None, reference_im=None, geom_col='geometry',
     for idx, feat in enumerate(feature_list):
         output_arr[:, :, idx] = features.rasterize([feat], out_shape=shape,
                                                    transform=affine_obj)
+
+    if reference_im:
+        reference_im = _check_rasterio_im_load(reference_im)
+    try:
+            bad_data_mask = (reference_im.read() == reference_im.nodata).any(axis=0) # take logical and along all dims so that all pixxels not -9999 across bands
+    except AttributeError as ae:  # raise another, more verbose AttributeError
+        raise AttributeError("A nodata value is not defined for the source image. Make sure the reference_im has a nodata value defined.") from ae
+        bad_data_mask = np.dstack([bad_data_mask]*output_arr.shape[2])
+        output_arr = np.where(bad_data_mask, 0, output_arr) # mask is broadcasted to filter labels where there are non-nan image values
+
     if out_file:
         meta = reference_im.meta.copy()
         meta.update(count=output_arr.shape[-1])